moved ZK related classes to common
