+++ twister2/common/src/java/BUILD
@@ -3,15 +3,21 @@ package(default_visibility = ["//visibility:public"])
 load("//tools:maven.bzl", "t2_java_lib")
 
 common_deps_files = [
-    "@org_yaml_snakeyaml//jar",
     "@com_google_protobuf_protobuf_java//jar",
     "@com_google_guava_guava//jar",
     "@com_esotericsoftware_kryo//jar",
+    "@org_apache_commons_commons_lang3//jar",
+    "@org_apache_curator_curator_client//jar",
+    "@org_apache_curator_curator_framework//jar",
+    "@org_apache_curator_curator_recipes//jar",
+    "@org_apache_zookeeper_zookeeper//jar",
+    "@org_yaml_snakeyaml//jar",
     "//twister2/proto:proto-java",
     "//twister2/api/src/java/edu/iu/dsc/tws/api/resource:resource-api-java",
     "//twister2/api/src/java/edu/iu/dsc/tws/api/comms:comms-api-java",
     "//twister2/api/src/java/edu/iu/dsc/tws/api/config:config-api-java",
     "//twister2/api/src/java/edu/iu/dsc/tws/api/exceptions:exceptions-java",
+    "//twister2/api/src/java/edu/iu/dsc/tws/api/faulttolerance:fault-tolerance-api-java",
     "//twister2/api/src/java/edu/iu/dsc/tws/api/net:network-api-java",
     "//twister2/api/src/java/edu/iu/dsc/tws/api/util:api-utils-java",
     "//third_party:ompi_javabinding_java",
